[286630] Add IClippingStrategy for selection handles
Selection handles are drawn on another layer than the selected figure.
So they are drawn also the corresponding editpart figure isn't.

Wrapping the selection handles in a figure allows to specifiy a
clipping strategy for multiple selections in different viewports.
But it is necessary to synchronize the bounds of the layer.